Disable INLINE_RUNTIME_CHUNK to avoid CSP issues
Chrome extensions can not execute inline scripts due to the Content
Security Policy. The CSP can be modified via content_security_policy
in manifest.json[1], but the restrictions can not be loosened enough to
allow for inline scripts to be executed.

The easy solution is to tell CRA not to inline it's bootstrap code, and
instead to serve it as an external script file[2].

[1] https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/content_security_policy
[2] https://stackoverflow.com/a/16153913